Add MyApp and HomeScreen widgets.
