+++ lib/widgets/home_screen/home_screen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class HomeScreen extends StatelessWidget {
+  const HomeScreen({Key key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Chat Messenger'),
+      ),
+      body: Center(
+        child: Text('Hello World'),
+      ),
+    );
+  }
+}
