+++ lib/widgets/my_app.dart
@@ -1,0 +1,11 @@
+import 'package:chat_messenger/widgets/home_screen/home_screen.dart';
+import 'package:flutter/material.dart';
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: HomeScreen(),
+    );
+  }
+}
